Float values are handled as doubles in the outstation.
Fixed something that was on my mind for a while.  A work around has been
added to JS in order to keep the peace.  I know the JSON standard does
not support inf, -inf and nan, but that is broken in the standard IMHO.

Sorry Node and Perl users, you willalso need to work around this for
analog values.  See my device.js code.